directories restructured to remove private _src file
