--- src/gwkokab/_src/parameters/__init__.py
@@ -1,43 +1,0 @@
-#  Copyright 2023 The GWKokab Authors
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-
-from .parameters import (
-    CHIRP_MASS as CHIRP_MASS,
-    COS_TILT_1 as COS_TILT_1,
-    COS_TILT_2 as COS_TILT_2,
-    DEFAULT_PRIORS as DEFAULT_PRIORS,
-    ECCENTRICITY as ECCENTRICITY,
-    EFFECTIVE_SPIN_MAGNITUDE as EFFECTIVE_SPIN_MAGNITUDE,
-    MASS_RATIO as MASS_RATIO,
-    ncopy as ncopy,
-    Parameter as Parameter,
-    PRIMARY_MASS_DETECTED as PRIMARY_MASS_DETECTED,
-    PRIMARY_MASS_SOURCE as PRIMARY_MASS_SOURCE,
-    PRIMARY_SPIN_MAGNITUDE as PRIMARY_SPIN_MAGNITUDE,
-    PRIMARY_SPIN_X as PRIMARY_SPIN_X,
-    PRIMARY_SPIN_Y as PRIMARY_SPIN_Y,
-    PRIMARY_SPIN_Z as PRIMARY_SPIN_Z,
-    REDSHIFT as REDSHIFT,
-    REDUCED_MASS as REDUCED_MASS,
-    SECONDARY_MASS_DETECTED as SECONDARY_MASS_DETECTED,
-    SECONDARY_MASS_SOURCE as SECONDARY_MASS_SOURCE,
-    SECONDARY_SPIN_MAGNITUDE as SECONDARY_SPIN_MAGNITUDE,
-    SECONDARY_SPIN_X as SECONDARY_SPIN_X,
-    SECONDARY_SPIN_Y as SECONDARY_SPIN_Y,
-    SECONDARY_SPIN_Z as SECONDARY_SPIN_Z,
-    SYMMETRIC_MASS_RATIO as SYMMETRIC_MASS_RATIO,
-    TILT_1 as TILT_1,
-    TILT_2 as TILT_2,
-)
