+++ src/gwkokab/debug/__init__.py
@@ -4,7 +4,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     https://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,22 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-ignore:
-  - "scripts/**"
 
-codecov:
-  require_ci_to_pass: true
-  branch: main
+from ._debug import debug, debug_flush, debug_mode, enable_debugging
 
-coverage:
-  precision: 2
-  range: 50..100
-  round: nearest
-  status:
-    project:
-      default:
-        target: auto
-        threshold: 1
-    patch:
-      default:
-        target: 90
+
+__all__ = ["debug", "debug_flush", "debug_mode", "enable_debugging"]
